+++ py3dtiles/exceptions.py
@@ -4,11 +4,9 @@ class Py3dtilesException(Exception):
 
     Client code that wishes to catch all py3dtiles exception can use `except Py3dtilesException`.
     """
-    pass
 
 
 class WorkerException(Py3dtilesException):
     """
     This exception will be thrown by the conversion code if one exception occurs inside a worker.
     """
-    pass