+++ tests/test_points.py
@@ -20,7 +20,7 @@ sample_points = np.array([[x / 30, x / 30, x / 30] for x in range(30)], dtype=np
 @pytest.fixture
 def node():
     bbox = np.array([[0, 0, 0], [2, 2, 2]])
-    return Node('noeud'.encode('utf-8'), bbox, compute_spacing(bbox))
+    return Node(b'noeud', bbox, compute_spacing(bbox))
 
 
 @pytest.fixture
@@ -78,24 +78,24 @@ def test_is_point_far_enough_perf(benchmark):
 
 
 def test_short_name_to_path():
-    short_tile_name = ''.encode("ascii")
+    short_tile_name = b''
     path = node_name_to_path(Path('work'), short_tile_name)
     assert str(path) == 'work/r'
 
 
 def test_long_name_to_path():
-    long_tile_name = '110542453782'.encode("ascii")
+    long_tile_name = b'110542453782'
     path = node_name_to_path(Path('work'), long_tile_name)
     assert str(path) == 'work/11054245/r3782'
 
 
 def test_long_name_to_path_with_extension():
-    long_tile_name = '110542453782'.encode("ascii")
+    long_tile_name = b'110542453782'
     path = node_name_to_path(Path('work'), long_tile_name, suffix='.pnts')
     assert str(path) == 'work/11054245/r3782.pnts'
 
 
 def test_long_name_to_path_with_short_split():
-    long_tile_name = '110542453782'.encode("ascii")
+    long_tile_name = b'110542453782'
     path = node_name_to_path(Path('work'), long_tile_name, split_len=2)
     assert str(path) == 'work/11/05/42/45/37/r82'